pyproject.toml
[tool.poetry]
name = "challenge-20220626"
version = "0.1.0"
description = ""
authors = ["gabriel.stoffel <gabriel.stoffel@gmail.com>"]
readme = "README.md"
packages = [{include = "challenge_20220626"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.3"
pymongo = "^4.4.1"
mongoengine = "^0.27.0"
scrapy-djangoitem = "^1.1.1"
django-environ = "^0.10.0"
django-rest-framework-mongoengine = "^3.4.1"
djangorestframework = "^3.14.0"
scrapy = "^2.9.0"
scrapy-mongoengine-item = "^0.1.5"
django-crontab = "^0.7.1"
apscheduler = "^3.10.1"
psutil = "^5.9.5"
schedule = "^1.2.0"
mongomock = "^4.1.2"
drf-yasg = "^1.21.6"
coreapi = "^2.3.3"
logging = "^0.4.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
